281024 2307 Checkpoint on memory allocations

+++ app/src/core/camera.c
@@ -2,7 +2,7 @@
 
 Camera2D camera;
 
-Camera2D create_camera(Vector2 position, u16 screenWidth, u16 screenHeight, f32 rotation) {
+Camera2D create_camera(Vector2 position, u16 screenWidth, u16 screenHeight, u8 rotation) {
     
     Camera2D cam;
