+++ app/src/core/fmath.c
@@ -3,7 +3,7 @@
 #include "math.h"
 #include "raymath.h"
 
-Vector2 get_a_point_of_a_circle(Vector2 position, f32 radius, f32 angle) {
+Vector2 get_a_point_of_a_circle(Vector2 position, i16 radius, i16 angle) {
     return (Vector2)
     {
         position.x + (radius * cos(angle * 3.1415f / 180.f)),
@@ -13,10 +13,10 @@ Vector2 get_a_point_of_a_circle(Vector2 position, f32 radius, f32 angle) {
 
 Vector2 move_towards(Vector2 position, Vector2 target, f32 speed) 
 {
-    return Vector2MoveTowards(position, target, speed);
+    return Vector2MoveTowards(position, target, (float)speed);
 }
 
-bool Vec2Equals(Vector2 v1, Vector2 v2, float tolerans) {
+bool Vec2Equals(Vector2 v1, Vector2 v2, i8 tolerans) {
 
     int tolerans1 = fabsf(v1.x - v2.x);
